Implement Security Filter (Without User Lookup)
Implemented the security filter. This is a filter which is applied
to all server requests. The filter will reject requests if the
requesting user does not posses a role appropriate for the attempted
action on the given endpoint.

This is only one layer of a depth in defence approach to protecting
the server's information assets. Simple role based security will not be
fine-grained enough for all security controls and so the individual
endpoint handlers will have to do some of the work themselves (such
as ensuring that only the users themselves have access to their
own chat messages, for instance).

In the future, the security filter will also be responsible for
fetching and validating user session data in order to verify the
identity and permissions of the requesting user and passing this data
down to the individual endpoint handlers.