Don't immediatley halt if cookie != auth header
It is possible under some legitmate use cases that a client could
submit an auth header and cookie auth token which have different
session IDs or user infos. Consequently, we can't immediatley halt the
request if this occurs because the client would then be stuck in limbo
without any means to recover (except perhaps waiting until all tokens
expire).

With this commit, we now treat requests where cookie != header as
aunauthorized but otherwise permit the request to continue (if the
resource does not require authentication). This way, the client
is able to recover simply by deleting the authentication token stored
in local storage.

Once JWT key storage is made persistent this error mode will become
less common but will still be possible and so a recovery option for
the client must always be present.