Add backend support for protest edit
Also reworked user objects to add a new unqiue user ID value (this
will break existing JWT tokens which will have to be deleted manually).

Purpose of the new field is to provide some protection in the case that
a user account is deleted and a new account created with the same
name. It is important that a deleted user's assets are not assumed by
the new account even if it is only possible in the brief window between
deletion of the user account and deletion or reclimation of the user's
assets (also consider that a server could crash in-between these
two steps).

This UUID should be used for all security sensitive operations on
user owned assets, rather than the username.