Remove the majority of JS formatting comments
It seems that they may be confusing webpack. Either that, or there
is some strange resource exhaustion bug in webpack which causes it
to fail to emit valid code and input file sizeis a contributing factor.

Note that this webpack bug was only encountered in development mode,
not production mode.