Remove note about mongo user requiring root
It appears that, despite the name, the dbAdmin role in Mongo doesn't
allow for anything too terribly desctructive. The role is required
for creating collections and is capable of profiling queries and
collecting stats.

At some point in the future, it may be worthwhile to have a seperate
user to setup the database and make collections just so as to follow
the principle of least privellege. However, for now, the main
application needs to create collections and there does not appear to
be any particular security risk associated with giving the application
user dbAdmin.