Create database version collection explicitly
Depending on the database config, the client will get an authorization
error when it attempts to read from a collection which does not exist.
Collections are implicitly created when an index is created or when
a document is inserted and so collections which have an index or which
are not read before they are written are okay.

For the LDPROTEST_DB_VERSION collection, there is no index and no
write attempt is made prior too the first read. Therefore, the
collection must be explicitly created.